requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
openpyxl==3.1.5
gunicorn==23.0.0
